refactor: move config to .env and extract database to separate module
