+++ .env
@@ -1,0 +1,8 @@
+# Bot Configuration
+BOT_TOKEN=
+
+# Admins 
+ADMINS=123,456,789
+
+# Log Channel
+LOG_CHANNEL=@test
