+++ requirements.txt
@@ -1,0 +1,2 @@
+aiogram>=3.0.0
+python-dotenv
