+++ src/database/__init__.py
@@ -1,0 +1,3 @@
+from .database import Database
+
+__all__ = ['Database']
